data-generator/requirements.txt
kafka-python==2.0.2
confluent-kafka==2.2.0
avro-python3==1.10.2
fastavro==1.8.0
requests==2.31.0
faker==19.6.2
pydantic==2.4.0
python-dotenv==1.0.0
schedule==1.2.0
numpy==1.24.3